requirements.txt
bitwarden-sdk
python-dotenv